requirements.txt
Flask==2.0.1
pandas==1.3.3
requests==2.26.0
apscheduler==3.8.1
gunicorn==20.1.0
python-dotenv==0.19.0